subida ok

+++ lib/src/api/categorias_api.dart
@@ -45,14 +45,13 @@ class CategoriasApi {
   final observacionesVariablesDatabase = ObservacionesVariablesDatabase();
 
   Future<bool> obtenerAmbos(BuildContext context) async {
-    print('ella no te ama');
     try {
       final url = '$_url/api/categoria/listar_categorias_productos';
+
+      final preferences = Preferences();
       final resp = await http.post(url, body: {});
       final Map<String, dynamic> decodedData = json.decode(resp.body);
       if (decodedData == null) return false;
-
-      final preferences = Preferences();
 
       // print(decodedData['result']['data'].length);
 
@@ -180,7 +179,7 @@ class CategoriasApi {
 
             int cantidadAdicionales = 0;
 
-            var cantiProAdi = productos[x]['producto_observaciones_fijas']; 
+            var cantiProAdi = productos[x]['producto_observaciones_fijas'];
 
             if (cantiProAdi['adicional_categoria']['opciones'].length > 0) {
               cantidadAdicionales++;
@@ -201,7 +200,8 @@ class CategoriasApi {
               cantidadAdicionales++;
             }
 
-            productosData.productoCantidadAdicional = cantidadAdicionales.toString();
+            productosData.productoCantidadAdicional =
+                cantidadAdicionales.toString();
 
             if (productosData.productoDestacado == '0') {
               productosData.productoEstadoDestacado = '0';
@@ -219,45 +219,72 @@ class CategoriasApi {
               productoDatabase.insertarProductosDb(productosData);
             }
 
-            await observacionesFijasDatabase.deleteObservacionesFijas(productos[x]['id_producto']);
-            await productosFijosDatabase.deleteProductosFijos(productos[x]['id_producto']);
-            await saboresDatabase.deleteSabores(productos[x]['id_producto']);
-            await opcionesSaboresDatabase.deleteOpcionesSabores(productos[x]['id_producto']);
+            if ( preferences.estadoCargaInicial == '1') {
 
-            await acompanhamientosDatabase.deleteAcompanhamientos(productos[x]['id_producto']);
-            await opcionesAcompanhamientosDatabase.deleteOpcionesAcompanhamientos(productos[x]['id_producto']);
+                  print('entro a esta  mierda');
+              await observacionesFijasDatabase
+                  .deleteObservacionesFijas(productos[x]['id_producto']);
+              await productosFijosDatabase
+                  .deleteProductosFijos(productos[x]['id_producto']);
+              await saboresDatabase.deleteSabores(productos[x]['id_producto']);
+              await opcionesSaboresDatabase
+                  .deleteOpcionesSabores(productos[x]['id_producto']);
 
-            await especialesADatabase.deleteEspecialesA(productos[x]['id_producto']);
-            await opcionesespecialesADatabase.deleteOpcionesEspecialesA(productos[x]['id_producto']);
+              await acompanhamientosDatabase
+                  .deleteAcompanhamientos(productos[x]['id_producto']);
+              await opcionesAcompanhamientosDatabase
+                  .deleteOpcionesAcompanhamientos(productos[x]['id_producto']);
 
-            await especialesBDatabase.deleteEspecialesB(productos[x]['id_producto']);
-            await opcionesespecialesBDatabase.deleteOpcionesEspecialesB(productos[x]['id_producto']);
+              await especialesADatabase
+                  .deleteEspecialesA(productos[x]['id_producto']);
+              await opcionesespecialesADatabase
+                  .deleteOpcionesEspecialesA(productos[x]['id_producto']);
 
-            await especialesCDatabase.deleteEspecialesC(productos[x]['id_producto']);
-            await opcionesespecialesCDatabase.deleteOpcionesEspecialesC(productos[x]['id_producto']);
+              await especialesBDatabase
+                  .deleteEspecialesB(productos[x]['id_producto']);
+              await opcionesespecialesBDatabase
+                  .deleteOpcionesEspecialesB(productos[x]['id_producto']);
 
-            await especialesDDatabase.deleteEspecialesD(productos[x]['id_producto']);
-            await opcionesespecialesDDatabase.deleteOpcionesEspecialesD(productos[x]['id_producto']);
+              await especialesCDatabase
+                  .deleteEspecialesC(productos[x]['id_producto']);
+              await opcionesespecialesCDatabase
+                  .deleteOpcionesEspecialesC(productos[x]['id_producto']);
 
-            await observacionesVariablesDatabase.deleteObservacionesVariables(productos[x]['id_producto']);
+              await especialesDDatabase
+                  .deleteEspecialesD(productos[x]['id_producto']);
+              await opcionesespecialesDDatabase
+                  .deleteOpcionesEspecialesD(productos[x]['id_producto']);
+
+              await observacionesVariablesDatabase
+                  .deleteObservacionesVariables(productos[x]['id_producto']);
+            }
 
             ObservacionesFijas observacionesFijas = ObservacionesFijas();
             observacionesFijas.idProducto = productos[x]['id_producto'];
-            observacionesFijas.mostrar =productos[x]['producto_observaciones_fijas']['mostrar_fijas'];
+            observacionesFijas.mostrar =
+                productos[x]['producto_observaciones_fijas']['mostrar_fijas'];
 
-            await observacionesFijasDatabase.insertarObservacionesFijas(observacionesFijas);
+            await observacionesFijasDatabase
+                .insertarObservacionesFijas(observacionesFijas);
 
-            var productillos =productos[x]['producto_observaciones_fijas']['productos'];
+            var productillos =
+                productos[x]['producto_observaciones_fijas']['productos'];
 
             if (productillos.length > 0) {
-              for (var z = 0;z <productos[x]['producto_observaciones_fijas']['productos'].length;z++) {
-                final productoIdbuscado = await productoDatabase.consultarPorId(productillos[z].toString());
+              for (var z = 0;
+                  z <
+                      productos[x]['producto_observaciones_fijas']['productos']
+                          .length;
+                  z++) {
+                final productoIdbuscado = await productoDatabase
+                    .consultarPorId(productillos[z].toString());
 
                 if (productoIdbuscado.length > 0) {
                   ProductosFijos productosFijos = ProductosFijos();
                   productosFijos.idProducto = productos[x]['id_producto'];
                   productosFijos.idRelacionado = productillos[z].toString();
-                  productosFijos.nombreProducto = productoIdbuscado[0].productoNombre;
+                  productosFijos.nombreProducto =
+                      productoIdbuscado[0].productoNombre;
 
                   await productosFijosDatabase
                       .insertarProductosFijos(productosFijos);
@@ -282,10 +309,13 @@ class CategoriasApi {
                   for (var t = 0; t < listOpcionesSabores.length; t++) {
                     OpcionesSabores opcionesSabores = OpcionesSabores();
                     opcionesSabores.idProducto = productos[x]['id_producto'];
-                    opcionesSabores.tituloTextos =saboresList[f]['titulo'].toString();
-                    opcionesSabores.nombreTexto =listOpcionesSabores[t].toString();
+                    opcionesSabores.tituloTextos =
+                        saboresList[f]['titulo'].toString();
+                    opcionesSabores.nombreTexto =
+                        listOpcionesSabores[t].toString();
 
-                    await opcionesSaboresDatabase.insertarOpcionesSabores(opcionesSabores);
+                    await opcionesSaboresDatabase
+                        .insertarOpcionesSabores(opcionesSabores);
                   }
                 }
               }
@@ -460,15 +490,22 @@ class CategoriasApi {
               print('ctm');
             }
 
-            var adicionalesList = productos[x]['producto_observaciones_fijas']['adicional_categoria']['opciones'];
-            var adicionalesList2 = productos[x]['producto_observaciones_fijas']['adicional_categoria_2']['opciones'];
-            var adicionalesList3 = productos[x]['producto_observaciones_fijas']['adicional_categoria_3']['opciones'];
-            var adicionalesList4 = productos[x]['producto_observaciones_fijas']['adicional_categoria_4']['opciones'];
-            var adicionalesList5 = productos[x]['producto_observaciones_fijas']['adicional_categoria_5']['opciones'];
-            var adicionalesList6 = productos[x]['producto_observaciones_fijas']['adicional_categoria_6']['opciones'];
+            var adicionalesList = productos[x]['producto_observaciones_fijas']
+                ['adicional_categoria']['opciones'];
+            var adicionalesList2 = productos[x]['producto_observaciones_fijas']
+                ['adicional_categoria_2']['opciones'];
+            var adicionalesList3 = productos[x]['producto_observaciones_fijas']
+                ['adicional_categoria_3']['opciones'];
+            var adicionalesList4 = productos[x]['producto_observaciones_fijas']
+                ['adicional_categoria_4']['opciones'];
+            var adicionalesList5 = productos[x]['producto_observaciones_fijas']
+                ['adicional_categoria_5']['opciones'];
+            var adicionalesList6 = productos[x]['producto_observaciones_fijas']
+                ['adicional_categoria_6']['opciones'];
 
             if (adicionalesList.length > 0) {
-              await adicionalesDatabase.deleteAdicionalesPorId(productos[x]['id_producto'], '0');
+              await adicionalesDatabase.deleteAdicionalesPorId(
+                  productos[x]['id_producto'], '0');
 
               for (var i = 0; i < adicionalesList.length; i++) {
                 AdicionalesModel adicionalesModel = AdicionalesModel();
@@ -476,7 +513,9 @@ class CategoriasApi {
                 adicionalesModel.idProducto = productos[x]['id_producto'];
                 adicionalesModel.idProductoAdicional = adicionalesList[i];
                 adicionalesModel.adicionalItem = '0';
-                adicionalesModel.titulo = productos[x]['producto_observaciones_fijas']['adicional_categoria']['titulo'];
+                adicionalesModel.titulo = productos[x]
+                        ['producto_observaciones_fijas']['adicional_categoria']
+                    ['titulo'];
                 adicionalesModel.adicionalSeleccionado = '0';
 
                 await adicionalesDatabase.insertarAdicionales(adicionalesModel);
@@ -594,5 +633,4 @@ class CategoriasApi {
       return false;
     }
   }
-
 }
